Удаляет предыдущие предложения на карте
